create component subfolders, move components, edit dependency paths
