+++ cryptonomicon-react/src/components/CryptoDetail/CryptoDetail.css
