+++ cryptonomicon-react/src/components/CryptoDetail/CryptoDetail.js
