+++ cryptonomicon-react/src/components/CryptoListItem/CryptoListItem.css
