+++ cryptonomicon-react/src/components/CryptoListItem/CryptoListItem.js
