+++ cryptonomicon-react/src/components/LoginSignup/Signup.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import adapter from '../adapter';
+import adapter from '../../adapter';
 import './LoginSignup.css';
 
 //Controlled component to handle Signup
