+++ cryptonomicon-react/src/components/Navbar/Navbar.js
